+++ packages/core/src/utils/getRedirectUriFromAuthConfig.ts
@@ -1,10 +1,15 @@
 import { Platform } from "react-native";
 
-export const getRedirectUriFromAuthConfig = (clientId: string, redirectUri: string, redirectUriWeb?: string): string => {
+export const getRedirectUriFromAuthConfig = (
+    redirectUri: string,
+    redirectUriWeb?: string,
+): string => {
     if (Platform.OS === "web") {
-        if (!redirectUriWeb) throw new Error("Looks like you started your application for web, but you have not registered a redirect URI for web. Please add a 'redirectUriWeb' to your mad config")
+        if (!redirectUriWeb)
+            throw new Error(
+                "Looks like you started your application for web, but you have not registered a redirect URI for web. Please add a 'redirectUriWeb' to your mad config",
+            );
         return redirectUriWeb;
     }
     return redirectUri;
-
-}
+};