+++ packages/dfw/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @equinor/mad-dfw
 
+## 0.0.16
+
+### Patch Changes
+
+-   Updated dependencies [49e643c]
+    -   @equinor/mad-components@0.13.0
+
+## 0.0.15
+
+### Patch Changes
+
+-   Updated dependencies [b5badbd]
+    -   @equinor/mad-components@0.12.10
+
+## 0.0.14
+
+### Patch Changes
+
+-   Updated dependencies [ba1d39f]
+    -   @equinor/mad-components@0.12.9
+
+## 0.0.13
+
+### Patch Changes
+
+-   5f69443: `PropertyRow` now correctly displays value when not in a `ScrollView` and on mobile
+-   Updated dependencies [8a09b94]
+    -   @equinor/mad-components@0.12.8
+
 ## 0.0.12
 
 ### Patch Changes